test(sn): keep DNS control regression in contract suite
